fix: Limpieza de código redundante y cambio de usar int a long para el ISBN

+++ src/Backend/TDA/Libros/NodoLibro.cs
@@ -2,18 +2,18 @@ using Backend.TDA.Categoria;
 
 namespace Backend.TDA.Libros
 {
-    public class NodoLibro
-    {
-        public int ISBN { get; set; }
-        public string Titulo { get; set; }
-        public string Autor { get; set; }
-        public NodoCategoria? Categoria { get; set; }
+	public class NodoLibro
+	{
+		public long ISBN { get; set; }
+		public string Titulo { get; set; }
+		public string Autor { get; set; }
+		public NodoCategoria? Categoria { get; set; }
 
-        public NodoLibro(int isbn, string titulo, string autor)
-        {
-            ISBN = isbn;
-            Titulo = titulo;
-            Autor = autor;
-        }
-    }
+		public NodoLibro(long isbn, string titulo, string autor)
+		{
+			ISBN = isbn;
+			Titulo = titulo;
+			Autor = autor;
+		}
+	}
 }
